svn path=/branches/ktorrent/dht/; revision=514270

+++ libktorrent/util/constants.h
@@ -23,12 +23,12 @@
 namespace bt
 {
 	typedef unsigned long long Uint64;
-	typedef unsigned long Uint32;
+	typedef unsigned int Uint32;
 	typedef unsigned short Uint16;
 	typedef unsigned char Uint8;
 
 	typedef long long Int64;
-	typedef long Int32;
+	typedef int Int32;
 	typedef short Int16;
 	typedef char Int8;
